requirements.txt
# Food Delivery Robot - Python Dependencies
# Cài đặt: pip install -r requirements.txt

# GUI Framework
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3

# Image Processing
Pillow==10.0.0

# Audio Support  
pygame==2.5.0

# HTTP Requests
requests==2.31.0

# JSON Processing
json5==0.9.14

# File System Operations
pathlib2==2.3.7

# Configuration Management
configparser==6.0.0

# Logging
colorlog==6.7.0

# Date/Time Utilities
python-dateutil==2.8.2

# ROS Communication (sẽ cần sau)
# roslibpy==1.5.0  # Uncomment when ready for ROS integration

# Development & Testing
pytest==7.4.2
pytest-qt==4.2.0

# Environment variables
python-dotenv==1.1.1

# Groq API
groq==0.4.1
# Pinned HTTP client for Groq SDK compatibility
httpx==0.27.2

# STT / TTS wrappers (đã dùng):
vosk==0.3.45

# Ghi âm micro
sounddevice==0.4.7
scipy==1.11.4

# Optional: Database support
# sqlite3 (built-in with Python)

# Optional: Serial Communication (for hardware)
# pyserial==3.5